+++ src/lib.rs
@@ -14,6 +14,7 @@ pub mod gb;
 pub mod gen;
 pub mod info;
 pub mod inst;
+pub mod license;
 pub mod macros;
 pub mod mmu;
 pub mod pad;